projects/glucose-prediction/requirements.txt
python-telegram-bot==20.3
pandas==1.5.3
numpy==1.23.5
joblib==1.2.0
lightgbm==3.3.5
xgboost==1.6.2
